refactor(client/src): review changes
