+++ packages/client/src/index.css
@@ -12,17 +12,12 @@ body {
   -moz-osx-font-smoothing: grayscale;
 }
 
-#root {
-  margin: 0;
-  width: 100%;
-  height: 100%;
-}
-
-root,
+.root,
 body,
 html {
   margin: 0;
   min-height: 100%;
+  width: 100%;
   height: 100%;
 }
 
@@ -33,17 +28,6 @@ h1 {
   margin: 30px auto;
   text-shadow: 2px 1px $dendy_button_primary;
   line-height: 1.5;
-}
-
-main {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  margin: auto;
-
-  @media (max-width: 730px) {
-    width: 95%;
-  }
 }
 
 .link {
