+++ packages/client/src/pages/Leaderboard/Leaderboard.css
@@ -19,10 +19,6 @@
     height: 70px;
   }
 
-  &__row {
-    height: 40px;
-  }
-
   &__cell-header {
     cursor: pointer;
   }
@@ -37,19 +33,19 @@
     transform: rotate(180deg);
   }
 
-  &__row:nth-child(even) {
+  &__row:nth-child(odd) {
     background-color: $background_color;
   }
 
-  &__row:nth-child(2) {
+  &__row:nth-child(1) {
     background-color: $dendy_button_primary;
   }
 
-  &__row:nth-child(3) {
+  &__row:nth-child(2) {
     background-color: $dendy_button_secondary;
   }
 
-  &__row:nth-child(4) {
+  &__row:nth-child(3) {
     background-color: $dendy_button_third;
   }
 }
