+++ packages/client/src/pages/Leaderboard/Leaderboard.test.tsx
@@ -3,7 +3,6 @@ import '@testing-library/jest-dom';
 import { render, screen } from '@testing-library/react';
 
 import { headerText, Leaderboard } from './Leaderboard';
-//for test update
 describe('Leaderboard', () => {
   test('it renders', () => {
     render(<Leaderboard header={headerText} />);
