+++ packages/client/src/pages/Leaderboard/LeaderboardRow/LeaderboardRow.css
@@ -1,0 +1,5 @@
+.leaderboard {
+  &__row {
+    height: 40px;
+  }
+}
