+++ packages/client/src/pages/Leaderboard/LeaderboardRow/typings.ts
@@ -6,4 +6,6 @@ type LeaderboardRowProps = {
   matches: number;
 };
 
-export type { LeaderboardRowProps };
+export type LeaderboardRowType = {
+  row: LeaderboardRowProps;
+};
