+++ packages/client/src/pages/Leaderboard/typing.ts
